refactor: split progress reporter into modular package
Split the monolithic _strategies.py into separate files for better
code organization:
- _strategy_protocol.py: Protocol definition
- _legacy_strategy.py: Legacy evaluation reporting strategy
- _coded_strategy.py: Coded evaluation reporting strategy
- _strategies.py: Re-exports for backward compatibility
